# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimVerificationStatus": {
        "enum": [
          "supported",
          "partially_supported",
          "unsupported",
          "contradictory"
        ],
        "title": "ClaimVerificationStatus",
        "type": "string"
      },
...
      "DocumentType": {
        "enum": [
          "resume",
          "cover_letter"
        ],
        "title": "DocumentType",
        "type": "string"
      },
...
      "ExportResponse": {
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "overflowed": {
            "default": false,
            "title": "Overflowed",
            "type": "boolean"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "pdf_available": {
            "title": "Pdf Available",
            "type": "boolean"
          }
        },
        "required": [
          "document_id",
          "pdf_available"
        ],
        "title": "ExportResponse",
        "type": "object"
      },
      "GenerateRequest": {
        "properties": {
          "document_type": {
            "$ref": "#/components/schemas/DocumentType",
            "default": "resume"
          },
          "mode": {
            "$ref": "#/components/schemas/TailoringMode",
            "default": "conservative"
          }
        },
        "title": "GenerateRequest",
        "type": "object"
      },
      "GenerateResponse": {
        "properties": {
          "generation_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Run Id"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "GenerateResponse",
        "type": "object"
      },
      "GeneratedClaimPublic": {
        "properties": {
          "claim_text": {
            "title": "Claim Text",
            "type": "string"
          },
          "claim_type": {
            "title": "Claim Type",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "evidence_repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Repository"
          },
          "evidence_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Title"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locator"
          },
          "verification_status": {
            "$ref": "#/components/schemas/ClaimVerificationStatus"
          }
        },
        "required": [
          "id",
          "claim_text",
          "claim_type",
          "verification_status",
          "confidence"
        ],
        "title": "GeneratedClaimPublic",
        "type": "object"
      },
      "GeneratedDocumentPublic": {
        "properties": {
          "content_json": {
            "additionalProperties": true,
            "title": "Content Json",
            "type": "object"
          },
          "generation_run_id": {
            "format": "uuid",
            "title": "Generation Run Id",
            "type": "string"
          },
          "html_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Ref"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "pdf_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Ref"
          },
          "plaintext_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plaintext Ref"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "job_id",
          "generation_run_id",
          "type",
          "template_id",
          "version",
          "content_json"
        ],
        "title": "GeneratedDocumentPublic",
        "type": "object"
      },
      "GenerationRunPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "estimated_cost": {
            "title": "Estimated Cost",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "prompt_versions": {
            "items": {
              "type": "string"
            },
            "title": "Prompt Versions",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/GenerationRunStatus"
          },
          "token_usage": {
            "additionalProperties": true,
            "title": "Token Usage",
            "type": "object"
          }
        },
        "required": [
          "id",
          "job_id",
          "pipeline_version",
          "status",
          "prompt_versions",
          "token_usage",
          "latency_ms",
          "estimated_cost",
          "created_at"
        ],
        "title": "GenerationRunPublic",
        "type": "object"
      },
      "GenerationRunStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "GenerationRunStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PositioningResponse": {
        "properties": {
          "gaps_to_avoid": {
            "items": {
              "type": "string"
            },
            "title": "Gaps To Avoid",
            "type": "array"
          },
          "project_priority": {
            "items": {
              "type": "string"
            },
            "title": "Project Priority",
            "type": "array"
          },
          "skills_priority": {
            "items": {
              "type": "string"
            },
            "title": "Skills Priority",
            "type": "array"
          },
          "target_identity": {
            "title": "Target Identity",
            "type": "string"
          },
          "themes": {
            "items": {
              "type": "string"
            },
            "title": "Themes",
            "type": "array"
          }
        },
        "required": [
          "target_identity",
          "themes",
          "project_priority",
          "skills_priority",
          "gaps_to_avoid"
        ],
        "title": "PositioningResponse",
        "type": "object"
      },
...
      "TailoringMode": {
        "enum": [
          "conservative",
          "full"
        ],
        "title": "TailoringMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Evidence-grounded AI career agent \u2014 API",
    "title": "ProofHire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedDocumentPublic"
                  },
                  "title": "Response List Documents Api V1 Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedDocumentPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/documents/{document_id}/claims": {
      "get": {
        "description": "Backs the resume workspace's provenance rail (PRD \u00a711.4): click a\nbullet, see exactly which repository/file it traces back to.",
        "operationId": "get_document_claims_api_v1_documents__document_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedClaimPublic"
                  },
                  "title": "Response Get Document Claims Api V1 Documents  Document Id  Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Claims",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/documents/{document_id}/export": {
      "post": {
        "description": "Synchronous (not queued): one bounded Playwright render + parse-back\nvalidation cycle, same rationale as resume upload and positioning.",
        "operationId": "export_document_endpoint_api_v1_documents__document_id__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ats_minimal",
                    "technical_dense",
                    "modern_editorial"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": "Letter",
              "enum": [
                "Letter",
                "A4"
              ],
              "title": "Page Size",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Document Endpoint",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/documents/{document_id}/pdf": {
      "get": {
        "operationId": "download_document_pdf_api_v1_documents__document_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Document Pdf",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/documents/{document_id}/preview": {
      "get": {
        "description": "Live preview for the resume workspace (checklist \u00a710.7).\n\nRuns the *same* template renderer the PDF export uses, minus the Chromium\nstep, so what the user sees while switching templates is what the export\nwill contain \u2014 not a separate React approximation that can drift from the\nreal output. Nothing is persisted; this is read-only.",
        "operationId": "preview_document_html_api_v1_documents__document_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ats_minimal",
                    "technical_dense",
                    "modern_editorial"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Document Html",
        "tags": [
          "generation"
        ]
      }
    },
...
    "/api/v1/generation-runs/{run_id}": {
      "get": {
        "operationId": "get_generation_run_api_v1_generation_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationRunPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Generation Run",
        "tags": [
          "generation"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}/generate": {
      "post": {
        "operationId": "trigger_generate_api_v1_jobs__job_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Generate",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/positioning": {
      "post": {
        "description": "Synchronous (not queued): one bounded LLM call, same rationale as resume\nupload. Not persisted \u2014 this is an ephemeral preview the user reviews\nbefore generating; POST /jobs/{id}/generate recomputes it internally as\npart of the same pipeline run, so nothing is lost by not storing it here.",
        "operationId": "get_positioning_api_v1_jobs__job_id__positioning_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositioningResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Positioning",
        "tags": [
          "generation"
        ]
      }
    },
...
  }
}